+++ flink-table/flink-table-runtime-blink/src/main/java/org/apache/flink/table/filesystem/stream/compact/CompactCoordinator.java
@@ -34,8 +34,8 @@ import org.apache.flink.streaming.api.operators.OneInputStreamOperator;
 import org.apache.flink.streaming.runtime.streamrecord.StreamRecord;
 import org.apache.flink.table.filesystem.stream.TaskTracker;
 import org.apache.flink.table.filesystem.stream.compact.CompactMessages.CompactionUnit;
-import org.apache.flink.table.filesystem.stream.compact.CompactMessages.CoordinatorOutput;
 import org.apache.flink.table.filesystem.stream.compact.CompactMessages.CoordinatorInput;
+import org.apache.flink.table.filesystem.stream.compact.CompactMessages.CoordinatorOutput;
 import org.apache.flink.table.filesystem.stream.compact.CompactMessages.EndCompaction;
 import org.apache.flink.table.filesystem.stream.compact.CompactMessages.EndInputFile;
 import org.apache.flink.table.filesystem.stream.compact.CompactMessages.InputFile;
